requirements.txt
adafruit-circuitpython-ads1x15
RPi.GPIO
urllib3
zipfile36
RPLCD
